Added most major functionality. The focus of this project has shifted.
I've come to realise that creating a whole new texture editor with the features required for it to be useful is a waste of time, and there is no need to reinvent the wheel. Instead, what this program does now is reload the image whenever a change is made on disk, and it displays a useful preview of how the tile will look in various positions and configurations. Changes should then be made with Paint.NET or Paint or Photoshop or whatever it is that you're using.